Reference

Contents

Reference

Contents

Index

Index

AppleHealthParser.AppleHealthParser

AppleHealthParser.attribute_meta

AppleHealthParser.basic_export

AppleHealthParser.enhanced_metadata

AppleHealthParser.flatten_dictionaries

AppleHealthParser.generate_csv

AppleHealthParser.return_child_elements

AppleHealthParser.return_child_elements_i

AppleHealthParser.AppleHealthParser — Module

Module for parsing Apple Health files.

AppleHealthParser.attribute_meta — Method

This function takes the vector of element attributes extracted using the XML parser and modifies it in place so that the parent elements are dictionaries containing keys and values from all the nested child elements appended as vectors.

AppleHealthParser.basic_export — Method

This function takes the original XML file downloaded from Apple and converts the data into separate CSV files according to the procedure outlined in the documentation.

AppleHealthParser.enhanced_metadata — Method

This function takes the original XML file downloaded from Apple and creates separate collections of dictionaries for each parent node. The metadata is preserved as key value pairs appended to the original dictionaries.

AppleHealthParser.flatten_dictionaries — Method

This function takes a dictionary with nested vectors containing child elements and flattens the data resulting in a dictionary that only has keys and values.

AppleHealthParser.generate_csv — Method

This function takes the element tags, element attributes and names of parent nodes and generates a csv file for each parent node name.

AppleHealthParser.return_child_elements — Method

This function returns the child element indices of the element attributes vector.

AppleHealthParser.return_child_elements_i — Method

This function returns two outputs. The first are the indices of child elements from the file which can be removed. The second output is the collection of indices which are parent elements for which child elements can be appended to the extracted dictionary.